Split parser into different files
This is the first step in order to be able to bundle all those parsers in their own file to get rid of all the dependencies that prettier has.
