+++ src/parser-flow.js
@@ -1,0 +1,26 @@
+"use strict";
+
+const createError = require("./parser-create-error");
+
+function parse(text) {
+  // Inline the require to avoid loading all the JS if we don't use it
+  const flowParser = require("flow-parser");
+
+  const ast = flowParser.parse(text, {
+    esproposal_class_instance_fields: true,
+    esproposal_class_static_fields: true,
+    esproposal_export_star_as: true
+  });
+
+  if (ast.errors.length > 0) {
+    throw createError(
+      ast.errors[0].message,
+      ast.errors[0].loc.start.line,
+      ast.errors[0].loc.start.column
+    );
+  }
+
+  return ast;
+}
+
+module.exports = parse;
